+++ src/routes.js
@@ -8,7 +8,7 @@ const Register = React.lazy(() => import("./componens/auth/register/Register"))
 const Cart = React.lazy(() => import("./componens/cart/Cart"))
 const AdminDashboard = React.lazy(() => import("./componens/admin/dashboard/AdminDashboard"))
 const AdminFileList = React.lazy(() => import("./componens/admin/files/AdminFileList"))
-const AdminCategories = React.lazy(() => import("./componens/admin/CategoryAndSubCategoryContainer.js"))
+const CategoryAndSubCategoryContainer =React.lazy(() => import("./componens/admin/categories/CategoryAndSubCategoryContainer"))
 
 const routes = [
   { path: "/", exact: true, name: "Home", component: mainContent },
@@ -17,7 +17,7 @@ const routes = [
   { path: "/cart", exact: true, name: "Cart", component: Cart },
   { path: "/admin", exact: true, name: "Cart", component: AdminDashboard },
   { path: "/admin/files", exact: true, name: "Cart", component: AdminFileList },
-  { path: "/admin/categories", exact: true, name: "Categories", component:AdminCategories }
+  { path: "/admin/categories", exact: true, name: "Categories", component:CategoryAndSubCategoryContainer }
 ];
 
 export default routes;
